backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pgvector==0.3.5
psycopg2-binary==2.9.9

# Redis
redis==5.1.0

# AI / LLM
anthropic==0.36.0
openai==1.51.0          # for text-embedding-3-small

# Google / Gmail
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0

# WhatsApp (Meta Cloud API — direct HTTP calls)
httpx==0.27.2

# Config / env
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Utilities
tenacity==9.0.0          # retry logic
structlog==24.4.0        # structured logging
python-dateutil==2.9.0

# Kafka (Stage 2 — event streaming)
aiokafka==0.11.0

# OpenAI Agents SDK (Stage 2 — production agent)
openai-agents==0.0.12    # pip install openai-agents

# MCP Server (Exercise 1.4)
mcp==1.0.0               # pip install mcp

# Load testing
locust==2.31.5           # pip install locust

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
aiosqlite==0.20.0        # in-memory SQLite for tests
httpx==0.27.2            # also used for test client
